+++ ChipmunkPro-iPhone-Trial/cpMarch.h
@@ -13,7 +13,7 @@ typedef void (*cpMarchSegmentFunc)(cpVect v0, cpVect v1, void *data);
 /// The given number of samples will be taken and spread across the bounding box area using the sampling function and context.
 /// The segment function will be called for each segment detected that lies along the density contour for @c threshold.
 void cpMarchSoft(
-  cpBB bb, int x_samples, int y_samples, cpFloat threshold,
+  cpBB bb, unsigned long x_samples, unsigned long y_samples, cpFloat threshold,
   cpMarchSegmentFunc segment, void *segment_data,
   cpMarchSampleFunc sample, void *sample_data
 );
@@ -22,7 +22,7 @@ void cpMarchSoft(
 /// The given number of samples will be taken and spread across the bounding box area using the sampling function and context.
 /// The segment function will be called for each segment detected that lies along the density contour for @c threshold.
 void cpMarchHard(
-  cpBB bb, int x_samples, int y_samples, cpFloat threshold,
+  cpBB bb, unsigned long x_samples, unsigned long y_samples, cpFloat threshold,
   cpMarchSegmentFunc segment, void *segment_data,
   cpMarchSampleFunc sample, void *sample_data
 );
